FastApi/background_remover/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
rembg
pillow==10.1.0
python-multipart==0.0.6
python-dotenv==1.0.0